LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/sticky/view-timeline-sticky-offscreen-2-expected.txt

Subject

FAIL View timeline bottom-sticky during entry and top-sticky during exit. assert_equals: Effect at the midpoint of the active range: cover 0% to cover 100% expected "0.5" but got "0.516667"
